+++ Gruntfile.js
@@ -1,7 +1,7 @@
 /**
- * Copyright 2014 LaxarJS
+ * Copyright 2015 LaxarJS
  * Released under the MIT license.
- * www.laxarjs.org
+ * http://laxarjs.org/license
  */
 /*global module,__dirname,__filename */
 module.exports = function( grunt ) {
@@ -26,6 +26,18 @@ module.exports = function( grunt ) {
             jshintrc: __dirname + '/.jshintrc'
          }
       },
+      cssmin: {
+         default: {
+            options: {
+               keepSpecialComments: 0
+            },
+            files: [{
+               expand: true,
+               src: 'var/static/css/*.theme.css',
+               ext: '.min.css'
+            }]
+         }
+      },
       compress: {
          default: {
             options: {
@@ -46,50 +58,62 @@ module.exports = function( grunt ) {
             } ]
          }
       },
-      portal_angular_dependencies: {
+      laxar_application_dependencies: {
          default: {
             options: {},
-            dest: 'var/static/portal_angular_dependencies.js',
+            dest: 'var/static/laxar_application_dependencies.js',
             src: [ 'application/flow/*.json' ]
          }
       },
-      css_merger: { default: {} },
-      widget_json_merger: { default: {} },
+      css_merger: {
+         default: {
+            src: [ 'application/flow/*.json' ]
+         }
+      },
       directory_tree: {
-         layouts: {
-            dest: 'var/listing/application_layouts.json',
+         application: {
+            dest: 'var/listing/application_resources.json',
             src: [
+               'application/flow/**/*.json',
                'application/layouts/**/*.css',
-               'application/layouts/**/*.html'
-            ]
+               'application/layouts/**/*.html',
+               'application/pages/**/*.json'
+            ],
+            options: {
+               embedContents: [
+                  'application/flow/**/*.json',
+                  'application/layouts/**/*.html',
+                  'application/pages/**/*.json'
+               ]
+            }
          },
-         themes: {
-            dest: 'var/listing/includes_themes.json',
-            src: [
-               'includes/themes/**/*.css',
-               'includes/themes/**/*.html'
-            ]
-         },
-         uikit_themes: {
-            dest: 'var/listing/laxar_uikit_themes.json',
+         bower_components: {
+            dest: 'var/listing/bower_components_resources.json',
             src: [
                'bower_components/laxar_uikit/themes/**/*.css',
-               'bower_components/laxar_uikit/themes/**/*.html'
-            ]
+               'bower_components/laxar_uikit/controls/**/*.+(css|html)'
+            ],
+            embedContents: [ 'bower_components/laxar_uikit/controls/**/*.html' ]
          },
-         uikit_controls: {
-            dest: 'var/listing/laxar_uikit_controls.json',
+         includes: {
+            dest: 'var/listing/includes_resources.json',
             src: [
-               'bower_components/laxar_uikit/controls/**/*.css',
-               'bower_components/laxar_uikit/controls/**/*.html'
-            ]
-         },
-         widgets: {
-            dest: 'var/listing/includes_widgets.json',
-            src: [
+               'includes/lib/laxar_uikit/themes/**/*.css',
+               'includes/lib/laxar_uikit/controls/**/*.+(css|html)',
+               'includes/themes/**/*.+(css|html)',
                'includes/widgets/*/*/*.+(css|html|json)',
-               'includes/widgets/*/*/!(bower_components|node_modules)/**/*.+(css|html|json)'
-            ]
+               '!includes/widgets/*/*/+(package|bower).json',
+               'includes/widgets/*/*/!(bower_components|node_modules|spec)/**/*.+(css|html|json)'
+            ],
+            options: {
+               embedContents: [
+                  'includes/lib/laxar_uikit/controls/**/*.html',
+                  'includes/themes/**/controls/**/*.html',
+                  'includes/widgets/*/*/widget.json',
+                  'includes/widgets/*/*/*.theme/*.html',
+                  'includes/widgets/*/*/*.theme/css/*.css'
+               ]
+            }
          }
       },
       requirejs: {
@@ -97,40 +121,45 @@ module.exports = function( grunt ) {
             options: {
                mainConfigFile: 'require_config.js',
                name: '../init',
-               out: 'var/build/optimized_init.js'
+               out: 'var/build/optimized_init.js',
+               optimize: 'uglify2'
             }
          }
       },
       watch: {
-         options:  {
-            livereload: liveReloadPort
+         options: {
+            livereload: liveReloadPort,
+            reload: true
          },
          Gruntfile: {
-            files: __filename,
-            options: {
-               /* reload Grunt config */
-               reload: true
-            }
+            files: __filename
          },
-         layouts: {
-            files: [ '<%= directory_tree.layouts.src %>' ],
-            tasks: [ 'directory_tree:layouts' ],
+         application: {
+            files: [
+               'application/**/!(scss)/*.*'
+            ]
+         },
+         libraries: {
+            files: [
+               'includes/lib/*/**',
+               'includes/controls/*/**',
+               'includes/themes/*.theme/**',
+               '!includes/**/(bower_components|node_modules)/**',
+               '!includes/**/*.scss'
+            ]
+         },
+         dependencies: {
+            files: [
+               '<%= directory_tree.application.src %>',
+               '<%= directory_tree.includes.src %>'
+            ],
+            tasks: [
+               'directory_tree:application',
+               'directory_tree:includes',
+               'portal_angular_dependencies'
+            ],
             options: {
                event: [ 'added', 'deleted' ]
-            }
-         },
-         themes: {
-            files: [ '<%= directory_tree.themes.src %>' ],
-            tasks: [ 'directory_tree:themes' ],
-            options: {
-               event: [ 'added', 'deleted' ]
-            }
-         },
-         widgets: {
-            files: [ '<%= directory_tree.widgets.src %>', 'includes/widgets/*' ],
-            tasks: [ 'directory_tree:widgets' ],
-            options: {
-               reload: true
             }
          }
       }
@@ -145,22 +174,25 @@ module.exports = function( grunt ) {
       var config = grunt.config( 'widget.' + widget );
       grunt.config( 'widget.' + widget, _.defaults( {}, config ) );
       grunt.config( 'watch.' + widget, {
-         files: [ widget + '/!(bower_components|node_modules)',
-                  widget + '/!(bower_components|node_modules)/**' ]/*,
-         tasks: [ 'widget:' + widget ]*/
+         files: [
+            widget + '/!(bower_components|node_modules)',
+            widget + '/!(bower_components|node_modules)/**',
+            '!' + widget + '/**/*.scss'
+         ]
       } );
    } );
 
    grunt.loadNpmTasks( 'grunt-laxar' );
+   grunt.loadNpmTasks( 'grunt-contrib-cssmin' );
    grunt.loadNpmTasks( 'grunt-contrib-compass' );
    grunt.loadNpmTasks( 'grunt-contrib-compress' );
    grunt.loadNpmTasks( 'grunt-contrib-watch' );
 
    grunt.registerTask( 'server', [ 'connect' ] );
-   grunt.registerTask( 'build', [ 'directory_tree', 'portal_angular_dependencies' ] );
-   grunt.registerTask( 'optimize', [ 'widget_json_merger', 'css_merger', 'requirejs' ] );
+   grunt.registerTask( 'build', [ 'directory_tree', 'laxar_application_dependencies' ] );
+   grunt.registerTask( 'optimize', [ 'build', 'css_merger', 'cssmin', 'requirejs' ] );
    grunt.registerTask( 'test', [ 'server', 'widgets' ] );
    grunt.registerTask( 'default', [ 'build', 'test' ] );
-   grunt.registerTask( 'dist', [ 'build', 'optimize', 'compress' ] );
+   grunt.registerTask( 'dist', [ 'optimize', 'compress' ] );
    grunt.registerTask( 'start', [ 'build', 'server', 'watch' ] );
 };