+++ require_config.js
@@ -1,38 +1,27 @@
 var require = {
    baseUrl: 'bower_components',
-   deps: [
-      'es5-shim/es5-shim',
-      'modernizr/modernizr'
-   ],
+   deps: [],
    shim: {
       angular: {
-         deps: [
-            'jquery'
-         ],
+         deps: [ 'jquery' ],
          exports: 'angular'
       },
       'angular-mocks': {
-         deps: [
-            'angular'
-         ],
+         deps: [ 'angular' ],
          init: function ( angular ) {
             'use strict';
             return angular.mock;
          }
       },
       'angular-route': {
-         deps: [
-            'angular'
-         ],
+         deps: [ 'angular' ],
          init: function ( angular ) {
             'use strict';
             return angular;
          }
       },
       'angular-sanitize': {
-         deps: [
-            'angular'
-         ],
+         deps: [ 'angular' ],
          init: function ( angular ) {
             'use strict';
             return angular;
@@ -41,18 +30,20 @@ var require = {
       'json-patch': {
          exports: 'jsonpatch'
       },
-      underscore: {
-         exports: '_',
-         init: function () {
-            'use strict';
-            return this._.noConflict();
-         }
-      },
+      //underscore: {
+      //   exports: '_',
+      //   init: function () {
+      //      'use strict';
+      //      return this._.noConflict();
+      //   }
+      //},
       handsontable: {
-          deps: [
-              'jquery', 'numeral'
-          ],
-         exports: 'Handsontable'
+         deps: [ 'jquery', 'numeral', 'pikaday/pikaday', 'css!pikaday/css/pikaday' ],
+         exports: 'Handsontable',
+         init: function(numeral, moment, pikaday) {
+            // Handsontable needs Pikaday in global scope.
+            this.Pikaday = pikaday;
+         }
       },
       d3: {
          exports: 'd3'
@@ -61,25 +52,18 @@ var require = {
          deps: [ 'd3' ],
          exports: 'nv'
       },
+      pikaday: {
+         exports: 'pikaday'
+      },
       'angular-nvd3': {
          deps: [ 'angular', 'nvd3' ]
       }
    },
    packages: [
       {
-         name: 'laxar',
-         location: 'laxar',
-         main: 'laxar'
-      },
-      {
-         name: 'laxar_patterns',
-         location: 'laxar_patterns',
-         main: 'laxar_patterns'
-      },
-      {
-         name: 'laxar_uikit',
-         location: 'laxar_uikit',
-         main: 'laxar_uikit'
+         name: 'laxar-application',
+         location: '..',
+         main: 'init'
       },
       {
          name: 'moment',
@@ -88,56 +72,65 @@ var require = {
       }
    ],
    paths: {
-      // LaxarJS Core:
+      // LaxarJS Core and dependencies:
+      laxar: 'laxar/dist/laxar.with-deps',
       requirejs: 'requirejs/require',
-      jquery: 'jquery/dist/jquery',
-      underscore: 'underscore/underscore',
+      text: 'requirejs-plugins/lib/text',
+      json: 'requirejs-plugins/src/json',
       angular: 'angular/angular',
       'angular-mocks': 'angular-mocks/angular-mocks',
       'angular-route': 'angular-route/angular-route',
       'angular-sanitize': 'angular-sanitize/angular-sanitize',
-      jjv: 'jjv/lib/jjv',
-      jjve: 'jjve/jjve',
 
-      // LaxarJS Core Testing:
+      //underscore: 'underscore/underscore',
+      //jjv: 'jjv/lib/jjv',
+      //jjve: 'jjve/jjve',
+
+      // LaxarJS Core (tests only):
+      'laxar/laxar_testing': 'laxar/dist/laxar_testing',
+      jquery: 'jquery/dist/jquery',
       jasmine: 'jasmine/lib/jasmine-core/jasmine',
       q_mock: 'q_mock/q',
 
-      // LaxarJS Core Legacy:
-      text: 'requirejs-plugins/lib/text',
-      json: 'requirejs-plugins/src/json',
-
       // LaxarJS Patterns:
+      'laxar-patterns': 'laxar-patterns/dist/laxar-patterns',
       'json-patch': 'fast-json-patch/src/json-patch-duplex',
 
       // LaxarJS UIKit:
-      jquery_ui: 'jquery_ui/ui',
-      'bootstrap-tooltip': 'bootstrap-sass-official/assets/javascripts/bootstrap/tooltip',
-      'bootstrap-affix': 'bootstrap-sass-official/assets/javascripts/bootstrap/affix',
-      trunk8: 'trunk8/trunk8',
+      'laxar-uikit': 'laxar-uikit/dist/laxar-uikit',
+      'laxar-uikit/controls': 'laxar-uikit/dist/controls',
 
-      // Spreadsheet-like table and dependencies
-      numeral: 'numeral/numeral',
-      handsontable: 'handsontable/dist/jquery.handsontable',
-
-      // Charts
-      d3: 'd3/d3',
-      nvd3: 'nvd3/nv.d3',
-      'angular-nvd3': 'angular-nvd3/dist/angular-nvd3',
-
-      // App Parts:
+      // LaxarJS application paths:
       'laxar-path-root': '..',
       'laxar-path-layouts': '../application/layouts',
       'laxar-path-pages': '../application/pages',
+      'laxar-path-flow': '../application/flow/flow.json',
       'laxar-path-widgets': '../includes/widgets',
       'laxar-path-themes': '../includes/themes',
-      'laxar-path-flow': '../application/flow/flow.json',
+      'laxar-path-default-theme': 'laxar-uikit/dist/themes/default.theme',
 
-      portal_angular_dependencies: '../var/static/portal_angular_dependencies'
+      // LaxarJS application modules (contents are generated):
+      'laxar-application-dependencies': '../var/static/laxar_application_dependencies',
+
+      // Spreadsheet-like table and dependencies
+      numeral: 'numeral/numeral',
+      handsontable: 'handsontable/dist/handsontable',
+
+      // Charts
+      d3: 'd3/d3',
+      nvd3: 'nvd3/build/nv.d3',
+      'angular-nvd3': 'angular-nvd3/dist/angular-nvd3',
+      pikaday: 'pikaday',
+
+      // App Parts:
+      'laxar-path-controls': '../includes/controls'
+
    },
    map: {
       '*': {
-         'css': 'require-css/css' // or whatever the path to require-css is
+         'css': 'require-css/css', // or whatever the path to require-css is
+         'laxar_uikit': 'laxar-uikit',
+         'laxar_patterns': 'laxar-patterns'
       }
    }
 };