+++ CHANGELOG.md
@@ -3,6 +3,24 @@
 ## Last Changes
 
 
+## v1.0.1
+
+- [#14](https://github.com/LaxarJS/mashup_demo/issues/14): Fixed `index.html` and refactored `debug.html` and `require.js`
+- [#13](https://github.com/LaxarJS/mashup_demo/issues/13): Removed test-results.xml of widgets
+
+
+## v1.0.0
+
+- [#12](https://github.com/LaxarJS/mashup_demo/issues/12): Added favicon
+- [#10](https://github.com/LaxarJS/mashup_demo/issues/10): Upgraded to LaxarJS major version and fixed breaking changes
+- [#11](https://github.com/LaxarJS/mashup_demo/issues/11): Fixed `resolutions` of bower configuration
+
+
+## v0.3.0
+
+- [#9](https://github.com/LaxarJS/mashup_demo/issues/9): theme: improved styling
+
+
 ## v0.2.0
 
 - [#8](https://github.com/LaxarJS/mashup_demo/issues/8): Added dependencies to the bower.json of each widget.